+++ devicelayer/src/DevicePcie.h
@@ -46,8 +46,8 @@ public:
   size_t getTraceBufferSizeMasterMinion(int device, TraceBufferType traceType) override;
   int updateFirmwareImage(int device, std::vector<unsigned char>& fwImage) override;
   int getDmaAlignment() const override;
-  uint64_t getDramSize() const override;
-  uint64_t getDramBaseAddress() const override;
+  uint64_t getDramSize(int device) const override;
+  uint64_t getDramBaseAddress(int device) const override;
   void* allocDmaBuffer(int device, size_t sizeInBytes, bool writeable) override;
   void freeDmaBuffer(void* dmaBuffer) override;
   DeviceConfig getDeviceConfig(int device) override;