+++ devicelayer/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] 
 ### Added
 ### Changed
+- *** Breaking change *** [SW-15492] Updated getDramBaseAddress() and getDramSize() to provide the device-specific information.
 ### Deprecated
 ### Removed
 ### Fixed